small cleanup

+++ src/PricerSolverWrappers.cpp
@@ -12,7 +12,7 @@ void NodeData::build_solve_mip() {
 
 void NodeData::construct_lp_sol_from_rmp() {
     lp_interface_get_nb_cols(RMP.get(), &nb_cols);
-    assert(nb_cols - id_pseudo_schedules == localColPool.size());
+    assert(nb_cols - id_pseudo_schedules == static_cast<int>(localColPool.size()));
 
     lambda.resize(localColPool.size(), 0.0);
     lp_interface_x(RMP.get(), lambda.data(), id_pseudo_schedules);
